+++ providers/dns/dns_providers.go
@@ -32,6 +32,7 @@ import (
 	"github.com/go-acme/lego/v3/providers/dns/dyn"
 	"github.com/go-acme/lego/v3/providers/dns/dynu"
 	"github.com/go-acme/lego/v3/providers/dns/easydns"
+	"github.com/go-acme/lego/v3/providers/dns/edgedns"
 	"github.com/go-acme/lego/v3/providers/dns/exec"
 	"github.com/go-acme/lego/v3/providers/dns/exoscale"
 	"github.com/go-acme/lego/v3/providers/dns/fastdns"
@@ -139,6 +140,8 @@ func NewDNSChallengeProviderByName(name string) (challenge.Provider, error) {
 		return dyn.NewDNSProvider()
 	case "dynu":
 		return dynu.NewDNSProvider()
+	case "edgedns":
+		return edgedns.NewDNSProvider()
 	case "fastdns":
 		return fastdns.NewDNSProvider()
 	case "easydns":